requirements.txt
pytorch-lightning==2.3.0
pattern==3.6
pytorch==2.3.0
transformers==4.28.0
nltk==3.8.1
python==3.9